[insert users to Mongo + refactoring]
