+++ db/objects/user.py
@@ -1,0 +1,7 @@
+class User:
+    user_ID = -1
+    name = ""
+
+    def __init__(self, user_ID, name, _id=None):
+        self.user_ID = user_ID
+        self.name = name
